Create Project 2 Main Code JavaScript
Added an undo button to the CSV Upload page. 

Had to re-engineer this process as the previous method made it impossible to import and remove information while updating a preview section at the users' discretion.  Instead of stripping the HTML and recreating the message every time a button is clicked, I instead hid and revealed the message/buttons when necessary to avoid calling extra functions and to avoid losing data along the way. 

This way, when an undo button is clicked, it hides the original message/buttons and reveals the appropriate message/buttons and the contact preview is updated. It was more straightforward and efficient as only two functions were used, instead of four.