Create Project 3 Main Code HTML
Changed the way a modal looks when asking the user if they want to add certain contacts as associated contacts to a company/deal.
Added their profile picture and email. If the contact does not have a profile picture, a graphic of their initials against a background of a randomized color is displayed instead. When the user clicks on their name/plus sign icon, the contact is added to the company/deal as an associated contact and taken off the list in the modal.